# Reactor hooks optimizations