@marcingrzejszczak 2021-01-05
Closing this since we're suggesting the manual reactor instrumentation mode.