@marcingrzejszczak 2021-01-12
You can wrap the calls to frameworks and logging (or will we provide proper abstractions e.g. in Gateway for better logging). Of course if you're interested you can improve the performance of the ON_EACH approach. Still, for now, we think that the best, suggested approach is to use the MANUAL instrumentation mode.